pyproject.toml
[project]
name = "tailortalk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "agent>=0.1.3",
    "dateparser>=1.2.1",
    "fastapi>=0.115.13",
    "google-api-python-client>=2.173.0",
    "google-auth>=2.40.3",
    "google-auth-oauthlib>=1.2.2",
    "groq>=0.28.0",
    "langchain-groq>=0.3.4",
    "langgraph>=0.4.10",
    "langsmith>=0.4.2",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "streamlit>=1.46.0",
]
